fixing compiles issues

+++ src/main/java/se/hoosierevents/project/springboot/config/UserRole.java
@@ -5,7 +5,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties("usertype")
 public interface UserRole {
 
-	final Integer USER_NORMAL = 1;
-	final Integer USER_ADMIN = 99;
-	final Integer USER_ORGANIZER = 2;
+	final Long USER_NORMAL = 1L;
+	final Long USER_ADMIN = 99L;
+	final Long USER_ORGANIZER = 2L;
 }
